test(bonding): add BondingSetup before tests
